+++ src/Call/Call.php
@@ -20,6 +20,7 @@ use Eloquent\Phony\Call\Event\ReturnedEventInterface;
 use Eloquent\Phony\Call\Event\ThrewEventInterface;
 use Eloquent\Phony\Call\Event\TraversableEventInterface;
 use Eloquent\Phony\Call\Exception\UndefinedCallException;
+use Eloquent\Phony\Call\Exception\UndefinedResponseException;
 use Eloquent\Phony\Collection\Exception\UndefinedIndexException;
 use Eloquent\Phony\Collection\IndexNormalizer;
 use Eloquent\Phony\Collection\IndexNormalizerInterface;
@@ -512,25 +513,54 @@ class Call implements CallInterface
     /**
      * Get the returned value.
      *
-     * @return mixed The returned value.
+     * @return mixed                      The returned value.
+     * @throws UndefinedResponseException If this call has not yet returned a value.
      */
     public function returnValue()
     {
         if ($this->responseEvent instanceof ReturnedEventInterface) {
             return $this->responseEvent->value();
         }
+
+        throw new UndefinedResponseException(
+            'The call has not yet returned a value.'
+        );
     }
 
     /**
      * Get the thrown exception.
      *
-     * @return Exception|Error|null The thrown exception, or null if no exception was thrown.
+     * @return Exception|Error            The thrown exception.
+     * @throws UndefinedResponseException If this call has not yet thrown an exception.
      */
     public function exception()
     {
         if ($this->endEvent instanceof ThrewEventInterface) {
             return $this->endEvent->exception();
         }
+
+        throw new UndefinedResponseException(
+            'The call has not yet thrown an exception.'
+        );
+    }
+
+    /**
+     * Get the response.
+     *
+     * @return tuple<Exception|Error|null,mixed> A 2-tuple of thrown exception or null, and return value.
+     * @throws UndefinedResponseException        If this call has not yet responded.
+     */
+    public function response()
+    {
+        if ($this->responseEvent instanceof ReturnedEventInterface) {
+            return array(null, $this->responseEvent->value());
+        }
+
+        if ($this->endEvent instanceof ThrewEventInterface) {
+            return array($this->endEvent->exception(), null);
+        }
+
+        throw new UndefinedResponseException('The call has not yet responded.');
     }
 
     /**