+++ src/Call/CallVerifier.php
@@ -24,6 +24,7 @@ use Eloquent\Phony\Call\Event\ReceivedExceptionEventInterface;
 use Eloquent\Phony\Call\Event\ResponseEventInterface;
 use Eloquent\Phony\Call\Event\TraversableEventInterface;
 use Eloquent\Phony\Call\Exception\UndefinedCallException;
+use Eloquent\Phony\Call\Exception\UndefinedResponseException;
 use Eloquent\Phony\Cardinality\Verification\AbstractCardinalityVerifier;
 use Eloquent\Phony\Event\EventCollectionInterface;
 use Eloquent\Phony\Event\EventInterface;
@@ -496,9 +497,10 @@ class CallVerifier extends AbstractCardinalityVerifier implements
     }
 
     /**
-     * Get the return value.
+     * Get the returned value.
      *
-     * @return mixed The return value.
+     * @return mixed                      The returned value.
+     * @throws UndefinedResponseException If this call has not yet returned a value.
      */
     public function returnValue()
     {
@@ -508,11 +510,23 @@ class CallVerifier extends AbstractCardinalityVerifier implements
     /**
      * Get the thrown exception.
      *
-     * @return Exception|Error|null The thrown exception, or null if no exception was thrown.
+     * @return Exception|Error            The thrown exception.
+     * @throws UndefinedResponseException If this call has not yet thrown an exception.
      */
     public function exception()
     {
         return $this->call->exception();
+    }
+
+    /**
+     * Get the response.
+     *
+     * @return tuple<Exception|Error|null,mixed> A 2-tuple of thrown exception or null, and return value.
+     * @throws UndefinedResponseException        If this call has not yet responded.
+     */
+    public function response()
+    {
+        return $this->call->response();
     }
 
     /**
@@ -748,13 +762,18 @@ class CallVerifier extends AbstractCardinalityVerifier implements
     {
         $cardinality = $this->resetCardinality()->assertSingular();
 
-        $responseEvent = $this->call->responseEvent();
-        $returnValue = $this->call->returnValue();
-        $exception = $this->call->exception();
+        if ($responseEvent = $this->call->responseEvent()) {
+            list($exception, $returnValue) = $this->call->response();
+
+            $hasReturned = !$exception;
+        } else {
+            $returnValue = null;
+            $hasReturned = false;
+        }
 
         if (0 === func_num_args()) {
             list($matchCount, $matchingEvents) =
-                $this->matchIf($responseEvent, $responseEvent && !$exception);
+                $this->matchIf($responseEvent, $hasReturned);
 
             if ($cardinality->matches($matchCount, 1)) {
                 return $this->assertionRecorder->createSuccess($matchingEvents);
@@ -767,7 +786,7 @@ class CallVerifier extends AbstractCardinalityVerifier implements
 
         list($matchCount, $matchingEvents) = $this->matchIf(
             $responseEvent,
-            $responseEvent && !$exception && $value->matches($returnValue)
+            $hasReturned && $value->matches($returnValue)
         );
 
         if ($cardinality->matches($matchCount, 1)) {
@@ -836,8 +855,12 @@ class CallVerifier extends AbstractCardinalityVerifier implements
     {
         $cardinality = $this->resetCardinality()->assertSingular();
 
-        $responseEvent = $this->call->responseEvent();
-        $exception = $this->call->exception();
+        if ($responseEvent = $this->call->responseEvent()) {
+            list($exception) = $this->call->response();
+        } else {
+            $exception = null;
+        }
+
         $isTypeSupported = false;
 
         if (null === $type) {